Add method to build `ServiceManager` class from builder (#1363)
